In HTTP-01, modularize the way of presenting the token.
First, with a simple local webserver.
