+++ acme/client.go
@@ -102,7 +102,7 @@ func NewClient(caURL string, usr User, keyBits int, optPort string) (*Client, er
 	// spec to this map. Otherwise they won`t be found.
 	solvers := make(map[string]solver)
 	solvers["simpleHttp"] = &simpleHTTPChallenge{jws: jws, optPort: optPort}
-	solvers["http-01"] = &httpChallenge{jws: jws, optPort: optPort}
+	solvers["http-01"] = &httpChallenge{jws: jws, method: &httpChallengeWebserver{optPort: optPort}}
 
 	return &Client{directory: dir, user: usr, jws: jws, keyBits: keyBits, solvers: solvers}, nil
 }
