+++ acme/http_challenge.go
@@ -4,25 +4,22 @@ import (
 	"encoding/json"
 	"errors"
 	"fmt"
-	"net"
 	"net/http"
-	"strings"
 	"time"
 )
 
+type httpChallengeMethod interface {
+	PresentToken(domain, token, keyAuth string, checkSolvedFunc func() error) (err error)
+}
+
 type httpChallenge struct {
-	jws     *jws
-	optPort string
-	start   chan net.Listener
-	end     chan error
+	jws    *jws
+	method httpChallengeMethod
 }
 
 func (s *httpChallenge) Solve(chlng challenge, domain string) error {
 
 	logf("[INFO] acme: Trying to solve HTTP-01")
-
-	s.start = make(chan net.Listener)
-	s.end = make(chan error)
 
 	// Generate the Key Authorization for the challenge
 	keyAuth, err := getKeyAuthorization(chlng.Token, &s.jws.privKey.PublicKey)
@@ -30,30 +27,20 @@ func (s *httpChallenge) Solve(chlng challenge, domain string) error {
 		return err
 	}
 
-	go s.startHTTPServer(domain, chlng.Token, keyAuth)
-	var listener net.Listener
-	select {
-	case listener = <-s.start:
-		break
-	case err := <-s.end:
-		return fmt.Errorf("Could not start HTTP server for challenge -> %v", err)
-	}
+	return s.method.PresentToken(domain, chlng.Token, keyAuth, func() error {
+		return s.checkSolved(chlng.URI, chlng.Type, chlng.Type, keyAuth)
+	})
+}
 
-	// Make sure we properly close the HTTP server before we return
-	defer func() {
-		listener.Close()
-		err = <-s.end
-		close(s.start)
-		close(s.end)
-	}()
+func (s *httpChallenge) checkSolved(cURI, cType, cToken, keyAuth string) error {
 
-	jsonBytes, err := json.Marshal(challenge{Resource: "challenge", Type: chlng.Type, Token: chlng.Token, KeyAuthorization: keyAuth})
+	jsonBytes, err := json.Marshal(challenge{Resource: "challenge", Type: cType, Token: cToken, KeyAuthorization: keyAuth})
 	if err != nil {
 		return errors.New("Failed to marshal network message...")
 	}
 
 	// Tell the server we handle HTTP-01
-	resp, err := s.jws.post(chlng.URI, jsonBytes)
+	resp, err := s.jws.post(cURI, jsonBytes)
 	if err != nil {
 		return fmt.Errorf("Failed to post JWS message. -> %v", err)
 	}
@@ -86,48 +73,8 @@ Loop:
 		}
 
 		time.Sleep(1 * time.Second)
-		resp, err = http.Get(chlng.URI)
+		resp, err = http.Get(cURI)
 	}
 
 	return nil
-}
-
-func (s *httpChallenge) startHTTPServer(domain string, token string, keyAuth string) {
-
-	// Allow for CLI port override
-	port := ":80"
-	if s.optPort != "" {
-		port = ":" + s.optPort
-	}
-
-	listener, err := net.Listen("tcp", domain+port)
-	if err != nil {
-		// if the domain:port bind failed, fall back to :port bind and try that instead.
-		listener, err = net.Listen("tcp", port)
-		if err != nil {
-			s.end <- err
-		}
-	}
-	// Signal successfull start
-	s.start <- listener
-
-	path := "/.well-known/acme-challenge/" + token
-
-	// The handler validates the HOST header and request type.
-	// For validation it then writes the token the server returned with the challenge
-	http.HandleFunc(path, func(w http.ResponseWriter, r *http.Request) {
-		if strings.HasPrefix(r.Host, domain) && r.Method == "GET" {
-			w.Header().Add("Content-Type", "text/plain")
-			w.Write([]byte(keyAuth))
-			logf("Served Key Authentication ...")
-		} else {
-			logf("Received request for domain %s with method %s", r.Host, r.Method)
-			w.Write([]byte("TEST"))
-		}
-	})
-
-	http.Serve(listener, nil)
-
-	// Signal that the server was shut down
-	s.end <- nil
 }
